server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
google-genai>=0.3.0
python-dotenv>=1.0.1

